Refactor PresetGrid.tsx and split sections into its own component
