+++ src/components/PresetGrid/FABStyles.ts
@@ -1,0 +1,91 @@
+import { StyleSheet } from 'react-native';
+
+/**
+ * Floating Action Button (FAB) styles
+ * Used for the main action button and its options menu
+ */
+export const fabStyles = StyleSheet.create({
+  // FAB Container
+  fabContainer: {
+    position: 'absolute',
+    bottom: 24,
+    right: 24,
+    alignItems: 'center',
+    zIndex: 1002,
+  },
+
+  // Main FAB
+  floatingButton: {
+    width: 56,
+    height: 56,
+    borderRadius: 28,
+    elevation: 8,
+    shadowOffset: { width: 0, height: 4 },
+    shadowOpacity: 0.3,
+    shadowRadius: 8,
+  },
+  mainFabButton: {
+    width: 56,
+    height: 56,
+    borderRadius: 28,
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+
+  // Mini FABs
+  miniFab: {
+    position: 'absolute',
+    right: 8,
+  },
+  miniFabButton: {
+    width: 40,
+    height: 40,
+    borderRadius: 20,
+    alignItems: 'center',
+    justifyContent: 'center',
+    elevation: 6,
+    shadowOffset: { width: 0, height: 3 },
+    shadowOpacity: 0.25,
+    shadowRadius: 6,
+  },
+
+  // FAB Overlay and Options Menu
+  fabOverlay: {
+    position: 'absolute',
+    top: 0,
+    left: 0,
+    right: 0,
+    bottom: 0,
+    backgroundColor: 'rgba(0, 0, 0, 0.5)',
+    justifyContent: 'center',
+    alignItems: 'center',
+    zIndex: 1001,
+  },
+  fabOptionsContainer: {
+    borderRadius: 12,
+    padding: 16,
+    minWidth: 200,
+    shadowColor: '#000',
+    shadowOffset: { width: 0, height: 4 },
+    shadowOpacity: 0.25,
+    shadowRadius: 12,
+    elevation: 12,
+  },
+  fabTitle: {
+    fontSize: 18,
+    fontWeight: 'bold',
+    marginBottom: 16,
+    textAlign: 'center',
+  },
+  fabOption: {
+    flexDirection: 'row',
+    alignItems: 'center',
+    paddingVertical: 16,
+    paddingHorizontal: 8,
+  },
+  fabOptionText: {
+    fontSize: 16,
+    fontWeight: '500',
+    marginLeft: 12,
+  },
+});
